+++ addOns/authhelper/CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Send the referer header on verification if set on the original request.
 - Removed requirement to set at least one header in the GUI for Header-Based Session Management.
 - Include step for errors in the authentication diagnostics.
+- Browser based authentication to also support HTTP basic authentication for Firefox.
+- Verification rule to improve detection.
 
 ### Fixed
 - Do not fail the authentication on diagnostic errors.